Refactor Question

+++ app/models/question.rb
@@ -13,10 +13,10 @@ class Question < ApplicationRecord
   scope :daily_questions, -> { where(created_at: (Time.now - 1.day)..Time.now) }
 
   def subscribe(user)
-    subscriptions.create(user: user)
+    subscriptions.create(user: user) if user
   end
 
   def unsubscribe(user)
-    subscriptions.where(user: user).first.destroy
+    subscriptions.where(user: user).first.destroy if user
   end
 end
